Use per source file folder compilation flags.
The flags retrieved from the compilation database, are merged together for each folder
containing a source file. This adds support for header files located in the same folder
as a source file, however it does not solve the missing flags for headers located in
folder which do not contain any source.

Also, this merges the compilation flags together, which is maybe not desired. Maybe a
configuration parameter to control this would be appropriate.